backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # PostgreSQL
pydantic
python-dotenv
openai
python-multipart
passlib[bcrypt]
bcrypt<4.1
python-jose[cryptography]
tika
python-magic
pytesseract>=0.3.10
Pillow>=9.0.0